+++ src/MEGASync/gui/TransferItem.h
@@ -23,8 +23,8 @@ public:
     void setSpeed(long long transferSpeed);
     void setTotalSize(long long size);
     void setType(int type, bool isSyncTransfer = false);
-    void setPriority(float priority);
-    float getPriority();
+    void setPriority(unsigned long long priority);
+    unsigned long long getPriority();
 
     void finishTransfer();
     void updateTransfer();
@@ -59,7 +59,7 @@ protected:
     long long totalSize, totalTransferredBytes;
     unsigned long long effectiveSpeed, effectiveTransferSpeed, lastUpdate, elapsedTransferTime;
     bool regular;
-    float priority;
+    unsigned long long priority;
 };
 
 Q_DECLARE_METATYPE(TransferItem*)
